refactor: threshold data-type to float and removed ability to update...

See merge request informatik/qe/swa_swe/swe/swess26/group5/g5t4!20

+++ src/main/java/at/qe/skeleton/models/ThresholdViolation.java
@@ -19,7 +19,7 @@ public class ThresholdViolation {
     private Metric metric;
 
     @Column(name = "`value`")
-    private Long value;
+    private Float value;
 
     @Enumerated(EnumType.STRING)
     private ViolationStatus violationStatus;
@@ -54,11 +54,11 @@ public class ThresholdViolation {
         this.metric = metric;
     }
 
-    public Long getValue() {
+    public Float getValue() {
         return value;
     }
 
-    public void setValue(Long value) {
+    public void setValue(Float value) {
         this.value = value;
     }
